Initial model of data movement.
This sets up a local memory for each worker, along with one shared
memory (stored in the policy domain). An ocrDataBlockList_t struct is
kept for each memory, to keep track of the total and used sizes, and the
DBs stored at that memory.

When a DB is created, it is stored in the worker's local memory.
When a DB is acquired, it is moved from its current location to the
current worker's local memory.
When a DB is released, nothing is done. It will be moved/evicted later,
when necessary.

If a worker's memory is too full, DBs are evicted to DRAM in LRU manner.
DRAM has no size limitations.

Change-Id: I64bbfdc1a44781a44fd0ecc5402e58faab1e1005