Another step towards OCR on FSIM.
This patch set adds a new structure for building OCR applicaitons.

It removes CRT from the applications. The minimum XE environment will
be setup by the kernel -- distribution of code & data sections, setup
of stack, setup of policy domain.

This removes (for the time being) stack overflow checking in order to
keep the complixity down. This likely will be added when the basics
are running and have been debugged with simple apps.

Change-Id: Ic6f2235eb89cd1e856915f415057eb24533fba97