OCR: Change ocrPolicyMsg_t to support in,out,in/out fields
This patch reworks the marshalling and unmarshalling of
ocrPolicyMsg_t structures across PDs by seperating fields into in,
out and in/out fields to reduce the amount of information that needs to
be marshalled and unmarshalled.

Other changes also include the automatic reset of ocrFatGuid_t fields when
crossing address spaces.

[See feature #225]
[See bug #273]
[See bug #284]

Change-Id: Ib3088cf43a445087b28e772b579285dd36dbdee2