Added support for a runtime-overhead profiler
The current profiler should only be used on x86 (can potentially
be ported using hardware counters). To activate, look at
the Makefile in x86-pthread-x86 and uncomment the relevant line.

When running a program, this will generate files called profiler_XXX
which should then be analyzed using the analyzeProfile.py script in
scripts/Profiler. The output will be gprof like. Refer to the README
file in scripts/Profiler for more detail.

Change-Id: Ia04784bcfa9174b980cfdf823a38668a2e98ea60