Allow EDTs to return Event GUIDs.
This fix allows an EDT to stand in as a proxy for other data-
producing machinery, kicking off the production of the data and
returning the future corresponding to the answer which will
(eventually) be calculated.

Add a test to exercise this case.  Without this patch, the test
dies with a segmentation fault, since depv[0].ptr is not valid.

Change-Id: Ia1ee89524ac66ab7868153acf1191db790d3c2b2