Improved SS and OCR build system and generalized FSim launch
The files are now built in the ./build/ directory and installed
in the ./install/ directory. This can be over-ridden however by setting
  - RMD_SRC (should point to the 'ss' directory containing the source)
  - RMD_INSTALL (should point to where you want things installed; this
    does not need to be in RMD_SRC)
You should launch the build from the directory you want to build in.
For example, if you want to build in /tmp/build, copy ss/build/* to
/tmp/build, go to /tmp/build and type 'make all'.

You can now launch FSim from anywhere and the configuration files allow
for environment variables in the paths to files. See
xe-sim/tests/nocr-printf/config.cfg for an example:
  - A new section called 'environment' allows you to define the
    default values of environment variables. You can then use them
    in any of the values of the configuration file as $(VAR) or ${VAR}
  - You can over-ride those default values by setting the environment
    variable when launching the FSim program

Modified OCR build of ocr-apps/printf to locally build all required
files (including the kernel which it will build a local copy of).
This new structure now allows you to point RMD_INSTALL to a directory
containing the toolchain, RMD_SRC to a directory containing the
source (your checkout for example) and build everything without
modifying the source of the installation. For example:
RMD_INSTALL=/tmp/install make -f Makefile.whole-fsim all install
will work from ocr/ocr-apps/printf (and use the source
in the same checked-out tree) and build in ocr-apps/printf/build
and install in ocr-apps/printf/install. You can then use:
/tmp/install/bin/fsim -s -c /tmp/install/fsim-configs/X -c config.cfg
from the install directory to run the application

Change-Id: Ic83e7b1bb93bb519cb0b6028318a6c4ef2b1d25f