OCR: Allow EDTs to acquire the same DB on multiple slots
This [fixes bug #106] where the EDT could not acquire
the same DB if it also used more than 64 slots.

You need to define OCR_MAX_MULTI_SLOT (by default 1) to
be the maximum number of 64 slots block that you need. This
is only required if an EDT tries to acquire the same DB on multiple
slots

Change-Id: I59b01cba9e6f691a370b2fd279dc4fbfc86d6735