Fix steal protocol for wst deque
Bug happens in the following scenario:

1- deque has >= 2 tasks
2- thief reads the tail
3- deque gets emptied (at least one pop)
4- thief reads the head
5- thief checks deque size (tail-head)
6- thief complete the steal

The check indicates the queue is not empty so the
thief proceed with cas-ing the head.
However, the deque is indeed emptied by step 3,
between the thief reading the tail and the head
(very small window I guess). So the thief reads
a stale value. As a consequence a task is executed
twice.

Credit goes to Zoran for pointing out the issue.

[closes bug #243]

Change-Id: I4625d7882ef6e78203627ad9e0df032d4e669e48