Sequential ID for neighbors
The policy-domain maintains a set of locations
for its neighbors. However, the global location
is not helpful for internal OCR modules like the
scheduler when indexing into different data structures
or reasoning about a sequential id of the neighbor.
So, when a policy-message is received the PD needs
to know which neighbor ID sent the message.
This neighbor ID is the sequential ID maintained
within this policy domain. Currently, the PD may
look up the srcLocation of the message on its
neighbor list and find out the index. That however
adds overhead on every message receive.
On the other hand, if the sender provides the
seqID in the policy message it does not need
to be looked up.

The sender now gets to know about that seqID after
it registers itself with the neighbor. In some case,
like TG, the receiver can simply extract the neighbor
seqId from the srcLocation. This however adds the
restriction that the neighbor list is always fixed
for TG and not up to experimentation.

[See bug #308]

Change-Id: I8e70b26c40e176df5ca2510d1608f7d9aa025797