Scheduler 1.0 Framework (xstack)
This changelist implements the new scheduler framework.
It also provides an example implementation of a heuristic
that uses this framework. This implementation is a
flat workstealing algorithm similar to the current HC scheduler
for running on shared memory. The new heuristic is disabled
by default. It can be enabled by switching on a flag called
SCHED_1_0 in the hc-policy.c file. The purpose of this
heuristic is to demonstrate a working heuristic that exercises
the new framework and its APIs.

[Fixes bug #75]

Change-Id: I8b69a14d0aa427d420f20c5b2301b0bd4bfae985