Cleaned up the switchRunlevel function in allocators
I made the first runlevel much clearer (the CONFIG_PARSE):
    - in phase 0, you need to update phasesPerRunlevel to contain the minimum
      number of phases YOU NEED (not provide but need)
    - in phase 1, you can, if needed, check that you can provide the requested
      number of phases.
The handling of that runlevel should be replicated across other modules

Most of the phase business currently deals with either GUIDs and memory. For the
GUID case, everyone doing a guidify needs to know that the GUID provider is up
and running before they do that (so they need 2 phases). In the memory case, TLSF
actually relies on someone initializing things and other allocators looking at
what that someone did (within a PD). There may still be issues with TLSF though.

Change-Id: Ic023bf9ad3339d74d1012134dd248b0d85c1d70b