Fix race condition on event destruction
- Fix issue where an event is satisfied in EDT A, which let the
runtime iterate and satisfy all its dependences. One of the dependence
may be an EDT B, that destroys the event. There's a race condition here
because the code that destroys the event did not set the DB_PROP_NO_RELEASE
flag on the event db free call. It means B would try to release that db
although it didn't acquire it. When A resumes, it may still try to read the
db waiter content and try to release it leading to all sorts of weird errors.

Change-Id: Id60a514082a305c614b141321f1d2fe435b5f8ff