Remove/fix include paths so TG-CE and its builder build again.
Oddly, the TG-CE target had an include path to the parallel Intel
repo's install directory and was picking up stddef.h from there. That
was a big problem because the stddef.h there came from Clang and was
not happy being compiled with gcc as part of OCR TG-CE...

Likewise, the TG-CE builder had bad paths.

Change-Id: I131595ce3413dd30f28bd1e0956e654f55ce07a7