OCR: TG and TG-x86 runlevels and comm-platform rework
The TG and TG-x86 implementations are now more similar and
TG-x86 is re-enabled by this patch.

It was necessary to rework the comm-platform for both platforms
because they were heavily involved in the shutdown protocol and
also made many assumptions that are now incorrect. The TG-x86
comm platform was totally rewritten and the TG one was slightly
modified. The mechanism for messaging was also changed. The
buffer for the response is now reserved prior to sending out
a message. This is to avoid deadlocks between the CEs.

Also updated handleless to use handles if they are already allocated.
Added a initHandle method for all commApi. This allows the caller to
initialize a pre-allocated handle (see above)

Bug changes:
    - Fixes #134 The shutdown protocol is much cleaner
    - Fixes #135 A more general location for TG/TG-x86 works
    - Fixes #557 TG runlevel implementation
    - Fixes #574 TG-x86 is back online
    - See #231 Neighbor identification was improved but not
      past 1 unit
    - See #527 Some leaks fixed
    - See #618 Some issues with CE communication were cleaned up
    - See #549 Runlevel support for FSim

Change-Id: I20dcfd390334227ef05bba47bfcb71bb2c55a1ad