Set ELS_USER_SIZE=4.
It was agreed today that ELS_USER_SIZE will be set to 4.
MPI-lite needs 3 slots today.  In the short term, all EDTs
will reserve 4 ELS slots whether they use MPI-lite or not.

In the long term, a flag will be added to indicate whether
an EDT wants to use ELS or not when it is created.

Change-Id: I43775c5016930a581af11a3e44b25a7894c1c98f