Another attempt to get legacy C running on real OCR.
The test app is hello.c in apps/legacy/x86_native.  This is where
I'm running the build of the app.

Method #5:
   1 - change the ocr/build/x86_lib/Makefile to strip all but the
       necessary symbols (for the test app)
   2 - do only one link (no partial link) to combine hello.o,
       newlib libs (crt0, liba, libm), and myocr.o.

The previous technique, partial linking, was producing bad executables
(not fully linked).  This technique seems to be better about producing
fully executable.  But it still has problems...  It looks as though the
.tdata and .tbss are not accessible from within the resulting hello
executable.

Change-Id: Ia4ab6333d6b13577b0da924c8999d75f57f22716