Sync OCR config files with apps/examples/*.cfg
The apps/examples/*/config.cfg files all specify 2MB sL2 memories for
the hardware but the software memory allocator is beign fed sizes for
the memory hierarchy from different OCR config files that were out of
sync (containing bigger size.)

This resulted in access beyond the end of sL2 during memory pool
initialization. Not sure why it wasn't crashing on cluster machines
(Fedora 20) but it was SEGV-ing on F23.

This fixes the symptom, but the long term fix should be to make the
FSIM & OCR config files auto-sync or auto-generate...

Change-Id: I058b3796068db93648304e8eb5dd2776dc952e83