OCR Make: build in a single prefix (no separate dir. per OCR_TYPE)
This will make it easier to build an OCR RPM and to use modules with OCR.

The major change is that everything now installs in a "standard" location
(bin/, lib/, share/) in OCR_INSTALL which now acts like a more traditional
prefix. To distinguish between OCR "flavors", OCR_TYPE is now appended
to the library built and to the ocrrun scripts. Note that include files
are installed once and are only removed with a squeaky make (an uninstall
no longer removes them because they are "shared" across OCR_TYPE). A similar
statement can be made on the scripts installed in share/ocr/scripts.

Change-Id: I38ae6e225f4b174581301efb6557a9a022cb0c08