OCR TG Communication: fixed race condition on XE alarm
A bug (nasty one) existed where the CE could process a message from
an XE while its alarm was still in flight. In some cases, this caused the
processing of the alarm (which has to read MSRs at a minimum) to read insane
values for those MSRs and could lead to a slew of problems including message
loss.

Change-Id: I5f13f418d7ce4a4d696d35379c9ff6da9f5f38db